requirements.txt
opencv-python-headless
numpy~=2.0.0
pillow~=10.3.0
requests~=2.32.01
PySide6
platformdirs